correcciones a bugs de arranque

+++ backend/src/main/java/ar/uba/fi/ingsoft1/todo_template/movies/MovieRatingDTO.java
@@ -1,6 +1,6 @@
 package ar.uba.fi.ingsoft1.todo_template.movies;
 
-public record MovieRatingDTO(long movieId,
-                             long userId,
-                             float rating) {
+public record MovieRatingDTO(Long movieId,
+                             Long userId,
+                             Long rating) {
 }
